[SPARK-43782][CORE] Support log level configuration with static Spark conf
### What changes were proposed in this pull request?

This PR proposes to add a static Spark conf to configure an override to the Log4j logging level. It’s a config version of the `SparkContext.setLogLevel()` semantics, when set it’ll trigger a log level override at the beginning of `SparkContext` startup.

### Why are the changes needed?

This is particularly helpful in the Spark Connect scenario where there’s no way for the client to call `setLogLevel` because `SparkContext` is not yet supported in its API, and when it connects to a platform where it can't change the Log4j config file.

### Does this PR introduce _any_ user-facing change?

Yes, it adds a new public configuration `"spark.log.level"`, and by default it's unset which means there's no override.

### How was this patch tested?

A new unit test in `SparkContextSuite`.

Closes #41302 from jwang0306/config-logging-level.

Authored-by: jwang0306 <jared.wang@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit e8f87413bdd905e5e208ac8484b3b11de1623fc2)